Rename some files and convert tabs to spaces
